# perf(native): use simd to speedup audio buffer mix