--- a/packages/frontend/native/media_capture/Cargo.toml
+++ b/packages/frontend/native/media_capture/Cargo.toml
@@ -6,6 +6,10 @@ version = "0.0.0"
 [lib]
 crate-type = ["cdylib", "rlib"]
 
+[[bench]]
+harness = false
+name    = "mix_audio_samples"
+
 [dependencies]
 napi            = { workspace = true, features = ["napi4"] }
 napi-derive     = { workspace = true, features = ["type-def"] }
@@ -23,6 +27,9 @@ objc2            = { workspace = true }
 objc2-foundation = { workspace = true }
 screencapturekit = { workspace = true }
 uuid             = { workspace = true, features = ["v4"] }
+
+[dev-dependencies]
+criterion2 = { workspace = true }
 
 [build-dependencies]
 napi-build = { workspace = true }